requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-meta==1.5.2
django-webpack-loader>=0.6.0
django-extensions==2.2.1
django-filer==1.5.0
django-polymorphic==2.0.3
blessings==1.7
curtsies==0.3.0
bpython==0.17.1
django-taggit==1.1.0
pandas==0.25.0
xlrd==1.2.0  # pandas extension
dateparser==0.7.1
environs==5.2.0
uwsgi==2.0.18
easy-thumbnails==2.6.0
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
djangorestframework==3.10.2
django-filter==2.2.0
django-rest-polymorphic
rules
celery
redis
pysolr
django-haystack
